dev: ee: Fix runtime env inside container for lpass, or if user has rvm, rbenv, or pyenv
General things that needed to be fixed in the Ansible Execution Environment
container:

LastPass CLI had issues with writing a socket file in the volume mounted $HOME,
and accessing this across the container/host boundary.  So, just create a
temporary directory for it.

If installed, tell rbenv or pyenv to use system Ruby or Python, respectively.

The problem with RVM is that it automatically sources .ruby-version,
.ruby-gemset, and .rvmrc inside the docker container.  However, the host's rvm
home and binaries may not be the same architecture nor compatible with the
container's OS.  RVM provides only a function to unload itself: __rvm_unload

So, we must invoke __rvm_unload in the right shell context.

This turns out to be at the end of ~/.bash_profile, which transits the volume
mount boundary to the host.

So, to avoid impacting the host's bash shells also, we set ANSIBLE_EE=1 and check
this first.

Also, the setup-ansible-ee.sh script is only copied into the docker container's
/etc/profile.d/ to avoid errant impacts on the host's bash shells that might be
invoked after.

POSIX shells should pick this up from there, meanwhile non-interactive sh and
bash scripts use $ENV and $BASH_ENV

Change-Id: I7141c174d8a402262455be8cc7949d51200734e4