dev: ee: Cleanup the .bash_profile shim after running non-interactive scripts too
When docker-run was passed another script or ansible-playbook command,
the .bash_profile shim wasn't being cleaned up.

Under normal conditions, the .bash_profile would be sourced on the next bash
shell that the user spawned, and cleaned up when that shell exited.
This was mostly harmless, but we still would rather clean up properly.

Change-Id: I2e9c64bed98c595d9a86ac2d2e8f9cc46d2a7f0e