dev: ee: Handle non-interactive bash shells where RVM is activated
For non-interactive shells, .bash_profile is not sourced, but instead .bashrc
is used. Rather than setup another self-cleaning oneliner there too, we note
that the place where RVM needs to be deactivated and matters most is just before
lastpass-ansible is called.

Therefore, we will do the check for ANSIBLE_EE=1 there and source the
in-container /etc/profile.d script to unload RVM just in time.

Change-Id: Ida3c1e4fe3f713ff72feff7dbb54b5eeb0fbb427